feat: CSD resize, DPI awareness, event queue pattern (v0.26.1)
* feat: GOGPU_DEBUG=1 env var enables DX12/Vulkan debug layer

* fix: update test mocks for wgpu HAL interface changes

- mockQueue: add SupportsCommandBufferCopies()
- mockTexture: add CurrentUsage()

* chore(deps): update wgpu v0.23.7, naga v0.16.4, gputypes v0.4.0 (v0.26.1)

- wgpu v0.23.7: GLES ADJUST_COORDINATE_SPACE, DX12 deferred resource destruction,
  DRED diagnostics, shader cache, Vulkan fence recycling
- naga v0.16.4: HLSL 72/72 parity, SPIR-V 164/164, ForceLoopBounding,
  zero-init loop (330x faster FXC), GLSL zero-init loop
- gputypes v0.4.0: new vertex format constants, blend constant type

* fix(wayland): defer CSD resize to after ack_configure (BUG-CSD-001)

ResizeCSD was called during xdg_toplevel.configure (phase 1), before
ack_configure. Compositor discarded subsurface state changes because
configure wasn't acknowledged yet.

Fix: store pending CSD dimensions in SetPendingCSDResize, apply in
xdgSurfaceConfigureCb after ack_configure. Subsurface state applies
atomically with parent commit. Matches GLFW pattern.

Needs WSL2 runtime testing: maximize, restore, close button.

* fix(wayland): CSD resize — GLFW destroy/recreate pattern, geometry fix (BUG-CSD-001)

- Maximize: destroy wl_surface + wl_subsurface (not just subsurface)
  to clear ghost pixels on WSLg
- Restore: recreate surface + subsurface from scratch
- set_window_geometry with CSD offset (-borderW, -titleBarH) so
  compositor knows actual window bounds for interactive resize
- Vulkan surface uses content size (minus CSD borders), not full
  configure size
- Skip expensive SHM recreate when buffer dimensions unchanged
  (position-only update for drag resize)

* feat(windows): programmatic DPI awareness — no manifest needed (TASK-GOGPU-DPI-001)

SetProcessDpiAwarenessContext(PerMonitorV2) with fallback to
SetProcessDPIAware. Fixes blurry text on 200%+ scale displays.
Users no longer need to create app.manifest + app.rc + windres.

* fix(windows): convert mouse coords physical→logical for high-DPI (BUG-GOGPU-DPI-001)

With DPI awareness enabled, WM_MOUSEMOVE and WM_POINTER report
physical pixels. UI layout uses logical coordinates (App.Size()).
Divide mouse X/Y by scale factor to match. Fixes mouse offset on
200%+ displays (@SideFx, @cdlus).

* refactor(wayland): event queue pattern + GLFW CSD geometry (BUG-WAYLAND-001)

- PollEvents: replace closeEmitted/hasResize/pendingWidth flags with
  event queue (same pattern as X11 and Windows platforms)
- CSD geometry: remove set_window_geometry in CSD mode (GLFW pattern).
  Compositor uses main surface bounds, configure = content size.
  Fixes interactive resize jump on first click.
- SetInitialGeometry removed (not needed with GLFW pattern)

* chore(deps): update wgpu v0.23.7 → v0.23.8

* fix(darwin): resolve 14 lint issues across macOS platform files

- errname: objcInitErr → errObjcInit
- gocritic: else{if} → else if
- govet: nolint unsafeptr for ObjC FFI pointers
- maintidx/nolintlint: fix stale nolint directives
- staticcheck: merge conditional into declaration
- unconvert: remove unnecessary NSUInteger casts
- whitespace: remove leading newline in closure

* chore(deps): update wgpu v0.23.7 → v0.23.9, naga v0.16.4 → v0.16.6

* style: gofmt platform_darwin.go, libwayland_csd.go

+++ internal/platform/darwin/events.go
@@ -109,14 +109,14 @@ func GetEventInfo(event ID) EventInfo {
 		NSEventTypeOtherMouseDown, NSEventTypeOtherMouseUp, NSEventTypeOtherMouseDragged,
 		NSEventTypeMouseMoved,
 		NSEventTypeTabletPoint:
-		info.Subtype = NSUInteger(event.GetUint64(selectors.subtype))
+		info.Subtype = event.GetUint64(selectors.subtype)
 		if info.Subtype == NSEventSubtypeTabletPoint || info.Type == NSEventTypeTabletPoint {
 			info.Pressure = event.GetDouble(selectors.pressure)
 			tilt := event.GetPoint(selectors.tilt)
 			info.TiltX = tilt.X
 			info.TiltY = tilt.Y
 			info.Rotation = event.GetDouble(selectors.rotation)
-			info.PointingDeviceType = NSUInteger(event.GetUint64(selectors.pointingDeviceType))
+			info.PointingDeviceType = event.GetUint64(selectors.pointingDeviceType)
 		}
 	}
